pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "db-migration-env"
version = "1.0.0"
description = "Stateful RL environment for database schema migration agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "openenv-core>=0.3.0",
    "requests>=2.31.0",
    "openai>=1.30.0",
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[tool.hatch.build.targets.wheel]
packages = ["."]
